==> 98. Data_Analysis copy.ipynb <==
import pandas as pd
import numpy as np
import os
import re
import time
import datetime
import gc
import string
import re
import seaborn as sns 
import matplotlib.pyplot as plt 
import math

from decimal import Decimal, getcontext

# --------------------------------------------

output_dir = r"E:\Cristian\1. Molecular Dinamics\2. Python Projects\3. Jupiter Code\7. Nose_Hoover Fixed\2. N_1000\1. Range_Ok\1. T_0.03"

# Construir las rutas de los archivos basadas en output_dir
particle_data_file = os.path.join(output_dir, "91. Sigma_&_Charge_MD.csv")
system_data_file = os.path.join(output_dir, "92. Potencial_MD.csv")
pos_data_file = os.path.join(output_dir, "Particle_Data.csv")

Sigma_DF = pd.read_csv(particle_data_file,sep='\t')
Potencial_DF = pd.read_csv(system_data_file,sep='\t')
Partic_DF = pd.read_csv(pos_data_file)


Partic_DF = Partic_DF[(Partic_DF['Step'] == Partic_DF['Step'].max())]

Potencial_DF['u/R'] = Potencial_DF['U']/4.0

Sigma_DF        = Sigma_DF[['u/R','Sigma','q/Q']]
Potencial_DF    = Potencial_DF[['U','u/R','Potencial_0.004', 'Potencial_0.04', 
                   'Potencial_0.08', 'Potencial_2.0', 'Potencial_4.0',
                   'Potencial_8.0', 'Potencial_40.0']]
Partic_DF       = Partic_DF[['Step','Particle','X_Pos','Y_Pos']]

Rou_Out = r'C:\Users\ESTACION\Desktop'

Sigma_DF.to_csv(os.path.join(Rou_Out,'Sigma_Data.csv'), sep='\t', index=False)
Potencial_DF.to_csv(os.path.join(Rou_Out,'Potencial_Data.csv'), sep='\t', index=False)
Partic_DF.to_csv(os.path.join(Rou_Out,'Positions.csv'), sep='\t', index=False)
